FIX enable accessway theme only for valid actions
Actions on the Security controller don't always return a themed response,
nor do we want the accessway theme unanimously applied for the entire site
as soon as the Security controller is pushed onto the controller stack.

To avoid this we can defer the setting of themes until after we've
identified that the Security controller will field the request directly,
and that the action is one that should have the theme applied.