FIX show Message and or Content when necessary
Previously important lead in text was being omitted from the final
render shown to the user. This included text about e.g. what to do
next, or why things haven't worked as expected.

silverstripe/mfa tests also failed when this module was also installed
as it depended on the output in $Message in order to pass a few tests.