FIX Unlock text from forcing horizontal scroll
On medium sized viewports (for all browsers) the viewport could be forced into a state
where it is neither tablet (which works fine) nor desktop (which also works fine), and
in this 'in-between' sizing the text would fill more than the width of the viewport,
causing a horizontal scrollbar to appear and deminish the user experience, through text
being less easily readable for sighted users.

Unfortuantely there is a discrepency with IE11 and CSS Grid support, where the interpretation
of the `auto` value is not as it is currently for modern browsers, and the work around was causing
this issue. The easy fix is to use a more approprate value for modern browsers, and let IE look
usable although not perfect. Which is acceptable.