DOCS update readme with steps to resovle requirements issues
When a project calls theme based front end assets via the Requirements
API this would fail out of the box with login forms, due to the theme
list being overwritten (thus no reference to the requested file could be
found). This is a bit of an edge case, and will vary from project to
project. The best resolution is to simply document this and provide
awareness on what action needs to happen (and the root cause) in order
to resolve this problem for any affected developers.